feat: add e2e testing and centralize cypress commands
